Added proxy support. Reduced cache_expire.
Added proxy support, so if using threadfin you can specify a proxy server to use and this script will honour it for both yt-dlp and all ffmpeg input streams. If proxy is not specified, it will go direct.

Reduced cache_expire units from days to minutes. Sometimes stream URL's are only valid for a short while. Default is 720 mins/12 hours.